% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{epoxy_style}
\alias{epoxy_style}
\alias{epoxy_style_apply}
\alias{epoxy_style_bold}
\alias{epoxy_style_code}
\alias{epoxy_style_collapse}
\alias{epoxy_style_get}
\alias{epoxy_style_html}
\alias{epoxy_style_inline}
\alias{epoxy_style_italic}
\alias{epoxy_style_set}
\alias{epoxy_style_wrap}
\title{Deprecated: \code{epoxy_style()}}
\usage{
epoxy_style(...)

epoxy_style_apply(...)

epoxy_style_bold(...)

epoxy_style_code(...)

epoxy_style_collapse(...)

epoxy_style_get(...)

epoxy_style_html(...)

epoxy_style_inline(...)

epoxy_style_italic(...)

epoxy_style_set(...)

epoxy_style_wrap(...)
}
\arguments{
\item{...}{Passed to the new \code{epoxy_transform} function.}
}
\description{
\code{epoxy_style()} was deprecated in favor of \code{\link[=epoxy_transform]{epoxy_transform()}} in \pkg{epoxy}
version 0.1.0.
}
\section{Functions}{
\itemize{
\item \code{epoxy_style()}: Deprecated in v0.1.0. Use \code{\link[=epoxy_transform]{epoxy_transform()}}
instead.

\item \code{epoxy_style_apply()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_apply]{epoxy_transform_apply()}} instead.

\item \code{epoxy_style_bold()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_bold]{epoxy_transform_bold()}} instead.

\item \code{epoxy_style_code()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_code]{epoxy_transform_code()}} instead.

\item \code{epoxy_style_collapse()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_collapse]{epoxy_transform_collapse()}} instead.

\item \code{epoxy_style_get()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_get]{epoxy_transform_get()}} instead.

\item \code{epoxy_style_html()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_html]{epoxy_transform_html()}} instead.

\item \code{epoxy_style_inline()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_inline]{epoxy_transform_inline()}} instead.

\item \code{epoxy_style_italic()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_italic]{epoxy_transform_italic()}} instead.

\item \code{epoxy_style_set()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_set]{epoxy_transform_set()}} instead.

\item \code{epoxy_style_wrap()}: Deprecated in \pkg{epoxy} v0.1.0. Use
\code{\link[=epoxy_transform_wrap]{epoxy_transform_wrap()}} instead.

}}
\keyword{internal}
