% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoxy_style_format.R
\name{epoxy_style_format}
\alias{epoxy_style_format}
\title{Style replacements with inline formatting}
\usage{
epoxy_style_format(
  bytes = scales::label_bytes(),
  date = scales::label_date(),
  date_short = scales::label_date_short(),
  time = scales::label_time(),
  dollar = scales::label_dollar(),
  log = scales::label_log(),
  number = scales::label_number(),
  comma = scales::label_comma(),
  number_auto = scales::label_number_auto(),
  ordinal = scales::label_ordinal(),
  parse = scales::label_parse(),
  math = scales::label_math(),
  percent = scales::label_percent(),
  pvalue = scales::label_pvalue(),
  scientific = scales::label_scientific(),
  wrap = scales::label_wrap(width = 80),
  transformer = glue::identity_transformer
)
}
\description{
Makes available all of the \code{label_} functions from the \pkg{scales} package
within an inline format function, \code{fmt()}. To apply formatting, wrap embraced
expressions in \code{fmt(expr, label)}, where \code{label} is the name of a labeller
function from \pkg{scales} (with or without the leading \code{label_}), e.g.
\code{"dollar"} or \code{"label_dollar"} for \code{\link[scales:label_dollar]{scales::label_dollar()}}. Short \code{label}
forms exist for some labellers, e.g. \code{"$"} for \code{"dollar"}.\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r\}
# Airbnb stats that may not be completely accurate
airbnb <- list(
 countries = 220,
 cities = 1e5,
 avg_stay = 4.326,
 avg_cost = 184.952,
 hosts = 4.12e6
)

# Define number/dollar styles to apply consistently
my_style <- epoxy_style_format(
 dollar = scales::label_dollar(accuracy = 1),
 number = scales::label_number(
   accuracy = 0.1,
   scale_cut = scales::cut_short_scale()
 )
)
```

```\{epoxy epoxy_style = my_style\}
* Airbnb includes listings in \{fmt(airbnb$cities, ",")\} cities
  in \{fmt(airbnb$countries, "auto")\} countries
  from around \{fmt(airbnb$hosts, "#")\} hosts.
  The average guest stays \{fmt(airbnb$avg_stay, "#")\} nights
  at a rate of \{fmt(airbnb$avg_cost, "$")\} per night.
```
}\if{html}{\out{</div>}}

With standard inline R code in R Markdown, we can write the following:\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{* Airbnb includes listings in 10<sup>5</sup> cities
  in 220 countries
  from around 4.12 &times; 10<sup>6</sup> hosts.
  The average guest stays 4.326 nights
  at a rate of 184.952 per night.
}\if{html}{\out{</div>}}
\itemize{
\item Airbnb includes listings in 10^5 cities
in 220 countries
from around 4.12 × 10^6 hosts.
The average guest stays 4.326 nights
at a rate of 184.952 per night.
}

Using \code{epoxy} and the light-weight \code{fmt()} function from \code{epoxy_style_format()},
we can improve the readability and formatting of the interwoven numbers.
\itemize{
\item Airbnb includes listings in 100,000 cities
in 220 countries
from around 4.1M hosts.
The average guest stays 4.3 nights
at a rate of $185 per night.
}
\subsection{\code{fmt()} labellers}{\tabular{ll}{
   \code{label} \tab Applies formatting with \cr
   \code{"bytes"} \tab \code{\link[scales:label_bytes]{scales::label_bytes()}} \cr
   \code{","}, \code{"comma"} \tab \code{\link[scales:label_number]{scales::label_comma()}} \cr
   \code{"d"}, \code{"date"} \tab \code{\link[scales:label_date]{scales::label_date()}} \cr
   \code{"date_short"}, \code{"ds"} \tab \code{\link[scales:label_date]{scales::label_date_short()}} \cr
   \code{"$"}, \code{"dollar"} \tab \code{\link[scales:label_dollar]{scales::label_dollar()}} \cr
   \code{"log"} \tab \code{\link[scales:label_log]{scales::label_log()}} \cr
   \code{"math"} \tab \code{\link[scales:label_parse]{scales::label_math()}} \cr
   \code{"#"}, \code{"number"} \tab \code{\link[scales:label_number]{scales::label_number()}} \cr
   \code{"a"}, \code{"auto"}, \code{"number_auto"} \tab \code{\link[scales:label_number_auto]{scales::label_number_auto()}} \cr
   \code{"o"}, \code{"ordinal"} \tab \code{\link[scales:label_ordinal]{scales::label_ordinal()}} \cr
   \code{"parse"} \tab \code{\link[scales:label_parse]{scales::label_parse()}} \cr
   \code{"\%"}, \code{"pct"}, \code{"percent"} \tab \code{\link[scales:label_percent]{scales::label_percent()}} \cr
   \code{"p"}, \code{"pvalue"} \tab \code{\link[scales:label_pvalue]{scales::label_pvalue()}} \cr
   \code{"scientific"}, \code{"si"} \tab \code{\link[scales:label_scientific]{scales::label_scientific()}} \cr
   \code{"dt"}, \code{"time"} \tab \code{\link[scales:label_date]{scales::label_time()}} \cr
   \code{"wrap"} \tab \code{\link[scales:label_wrap]{scales::label_wrap()}} \cr
}

}
}
