% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderEpoxyHTML}
\alias{renderEpoxyHTML}
\title{Render Epoxy Output}
\usage{
renderEpoxyHTML(
  ...,
  .list = NULL,
  env = parent.frame(),
  outputFunc = epoxyHTML,
  outputArgs = list()
)
}
\arguments{
\item{...}{Named values corresponding to the template variables created with
the associated \code{\link[=epoxyHTML]{epoxyHTML()}} UI element.}

\item{.list}{A named list or a \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list with names
corresponding to the template variables created with the associated
\code{\link[=epoxyHTML]{epoxyHTML()}} UI element.}

\item{env}{The environment in which to evaluate the \code{...}}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[=epoxyHTML]{epoxyHTML()}} when \code{renderEpoxyHTML} is used in an interactive R
Markdown document.}
}
\value{
A server-side Shiny render function that should be assigned to
Shiny's \code{output} object and named to match the \code{.id} of the corresponding
\code{\link[=epoxyHTML]{epoxyHTML()}} call.
}
\description{
Server-side render function used to provide values for template items. Use
named values matching the template variable names in the associated
\code{epoxyHTML()}. When the values are updated by the app, \code{renderEpoxyHTML()}
will update the values shown in the app's UI.
}
\examples{
# This small app shows the current time using `epoxyHTML()`
# to provide the HTML template and `renderEpoxyHTML()` to
# update the current time every second.

ui <- shiny::fluidPage(
  shiny::h2("Current Time"),
  epoxyHTML(
    "time",
    shiny::p("The current time is {{strong time}}.")
  )
)

server <- function(input, output, session) {
  current_time <- shiny::reactive({
    shiny::invalidateLater(1000)
    strftime(Sys.time(), "\%F \%T")
  })

  output$time <- renderEpoxyHTML(time = current_time())
}

if (interactive()) {
  shiny::shinyApp(ui, server)
}

}
\seealso{
\code{\link[=epoxyHTML]{epoxyHTML()}}
}
