% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoxy-package.R
\docType{package}
\name{epoxy-package}
\alias{epoxy}
\alias{epoxy-package}
\title{epoxy: String Interpolation for 'R Markdown' Documents}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extra strength 'glue' for data-driven templating. String interpolation for 'R Markdown' documents and 'Shiny' apps, built on the 'glue' and 'whisker' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/epoxy}
  \item Report bugs at \url{https://github.com/gadenbuie/epoxy/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
