% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoxy-package.R
\docType{package}
\name{epoxy-package}
\alias{epoxy}
\alias{epoxy-package}
\title{epoxy: String Interpolation for 'R Markdown' Documents}
\description{
Experimental. String interpolation for 'R Markdown'
    documents and 'Shiny' apps.
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
