% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{epoxy_style}
\alias{epoxy_style}
\alias{epoxy_style_wrap}
\alias{epoxy_style_bold}
\alias{epoxy_style_italic}
\alias{epoxy_style_code}
\alias{epoxy_style_collapse}
\title{epoxy Style Transformers}
\usage{
epoxy_style_wrap(before = "**", after = "**")

epoxy_style_bold()

epoxy_style_italic()

epoxy_style_code()

epoxy_style_collapse(
  sep = ", ",
  last = "",
  last_and = " and ",
  last_or = " or ",
  sep_and = sep,
  sep_or = sep
)
}
\arguments{
\item{before, after}{In \code{epoxy_style_wrap()}, the characters to be added
before and after variables in the template string.}

\item{sep, sep_and, sep_or}{The separator to use when joining the vector
elements when the variable ends in \code{*}, \code{&}, or \code{|} respectively. By
default, these are all \code{", "}.}

\item{last, last_and, last_or}{Additional text added after \code{sep} before the
last element when the variable ends in \code{*}, \code{&}, or \code{|} respectively.}
}
\description{
These transformers provide additional automatic formatting for the template
strings. They are designed to be used with the \code{.transformer} chunk option of
in \code{epoxy} chunks.
}
\section{Functions}{
\itemize{
\item \code{epoxy_style_wrap}: Wrap variables

\item \code{epoxy_style_bold}: Embolden variables using markdown \verb{**} syntax

\item \code{epoxy_style_italic}: Italicize variables using markdown \verb{_} syntax

\item \code{epoxy_style_code}: Code format variables using markdown backtick syntax

\item \code{epoxy_style_collapse}: Collapse vector variables.
}}

