% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{epoxy_style}
\alias{epoxy_style}
\alias{epoxy_style_apply}
\alias{epoxy_style_bold}
\alias{epoxy_style_code}
\alias{epoxy_style_collapse}
\alias{epoxy_style_get}
\alias{epoxy_style_html}
\alias{epoxy_style_inline}
\alias{epoxy_style_italic}
\alias{epoxy_style_set}
\alias{epoxy_style_wrap}
\title{Deprecated: \code{epoxy_style()}}
\usage{
epoxy_style(...)

epoxy_style_apply(...)

epoxy_style_bold(...)

epoxy_style_code(...)

epoxy_style_collapse(...)

epoxy_style_get(...)

epoxy_style_html(...)

epoxy_style_inline(...)

epoxy_style_italic(...)

epoxy_style_set(...)

epoxy_style_wrap(...)
}
\arguments{
\item{...}{Passed to the new \code{epoxy_transform} function.}
}
\description{
\code{epoxy_style()} was renamed \code{\link[=epoxy_transform]{epoxy_transform()}} in \pkg{epoxy} version 0.1.0.
}
\section{Functions}{
\itemize{
\item \code{epoxy_style()}: Renamed \code{\link[=epoxy_transform]{epoxy_transform()}}.

\item \code{epoxy_style_apply()}: Renamed \code{\link[=epoxy_transform_apply]{epoxy_transform_apply()}}.

\item \code{epoxy_style_bold()}: Renamed \code{\link[=epoxy_transform_bold]{epoxy_transform_bold()}}.

\item \code{epoxy_style_code()}: Renamed \code{\link[=epoxy_transform_code]{epoxy_transform_code()}}.

\item \code{epoxy_style_collapse()}: Renamed \code{\link[=epoxy_transform_collapse]{epoxy_transform_collapse()}}.

\item \code{epoxy_style_get()}: Renamed \code{\link[=epoxy_transform_get]{epoxy_transform_get()}}.

\item \code{epoxy_style_html()}: Renamed \code{\link[=epoxy_transform_html]{epoxy_transform_html()}}.

\item \code{epoxy_style_inline()}: Renamed \code{\link[=epoxy_transform_inline]{epoxy_transform_inline()}}.

\item \code{epoxy_style_italic()}: Renamed \code{\link[=epoxy_transform_italic]{epoxy_transform_italic()}}.

\item \code{epoxy_style_set()}: Renamed \code{\link[=epoxy_transform_set]{epoxy_transform_set()}}.

\item \code{epoxy_style_wrap()}: Renamed \code{\link[=epoxy_transform_wrap]{epoxy_transform_wrap()}}.

}}
\keyword{internal}
