% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{epoxy_style}
\alias{epoxy_style}
\alias{epoxy_style_wrap}
\alias{epoxy_style_bold}
\alias{epoxy_style_italic}
\alias{epoxy_style_code}
\alias{epoxy_style_collapse}
\title{epoxy Style Transformers}
\usage{
epoxy_style(...)

epoxy_style_wrap(
  before = "**",
  after = "**",
  transformer = glue::identity_transformer
)

epoxy_style_bold(transformer = glue::identity_transformer)

epoxy_style_italic(transformer = glue::identity_transformer)

epoxy_style_code(transformer = glue::identity_transformer)

epoxy_style_collapse(
  sep = ", ",
  last = "",
  last_and = " and ",
  last_or = " or ",
  sep_and = sep,
  sep_or = sep,
  transformer = glue::identity_transformer
)
}
\arguments{
\item{...}{A list of style functions, e.g. \code{epoxy_style_bold} or the name of
a style function, e.g. \code{"bold"}, or a call to a style function, e.g.
\code{epoxy_style_bold()}. \code{epoxy_style()} chains the style functions together,
applying the styles from left to right.

For example, \code{epoxy_style("bold", "collapse")} results in replaced strings
that are emboldened \emph{and then} collapsed, e.g. \verb{**a** and **b**}. On the
other hand, \code{epoxy_style("collapse", "bold")}  will collapse the vector
\emph{and then} embolden the entire string.}

\item{before, after}{In \code{epoxy_style_wrap()}, the characters to be added
before and after variables in the template string.}

\item{transformer}{The transformer to apply to the replacement string. This
argument is used for chaining the transformer functions. By providing a
function to this argument you can apply an additional transformation after
the current transformation. In nearly all cases, you can let
\code{epoxy_style()} handle this for you. The chain ends when
\code{\link[glue:identity_transformer]{glue::identity_transformer()}} is used as the \code{transformer}.}

\item{sep, sep_and, sep_or}{The separator to use when joining the vector
elements when the variable ends in \code{*}, \code{&}, or \code{|} respectively. By
default, these are all \code{", "}.}

\item{last, last_and, last_or}{Additional text added after \code{sep} before the
last element when the variable ends in \code{*}, \code{&}, or \code{|} respectively.}
}
\description{
These transformers provide additional automatic formatting for the template
strings. They are designed to be used with the \code{.transformer} chunk option of
in \code{epoxy} chunks. You can use \code{epoxy_style()} to chain several transformers
together.
}
\section{Functions}{
\itemize{
\item \code{epoxy_style_wrap}: Wrap variables

\item \code{epoxy_style_bold}: Embolden variables using markdown \verb{**} syntax

\item \code{epoxy_style_italic}: Italicize variables using markdown \verb{_} syntax

\item \code{epoxy_style_code}: Code format variables using markdown backtick syntax

\item \code{epoxy_style_collapse}: Collapse vector variables.
}}

\examples{
glue::glue("{letters[1:3]&}", .transformer = epoxy_style("bold", "collapse"))
glue::glue("{letters[1:3]&}", .transformer = epoxy_style("collapse", "bold"))

}
