% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{epoxy_style}
\alias{epoxy_style}
\alias{epoxy_style_wrap}
\alias{epoxy_style_bold}
\alias{epoxy_style_italic}
\alias{epoxy_style_apply}
\alias{epoxy_style_code}
\alias{epoxy_style_collapse}
\title{epoxy Style Transformers}
\usage{
epoxy_style(..., syntax = NULL)

epoxy_style_wrap(
  before = "**",
  after = before,
  syntax = NULL,
  transformer = glue::identity_transformer
)

epoxy_style_bold(syntax = NULL, transformer = glue::identity_transformer)

epoxy_style_italic(syntax = NULL, transformer = glue::identity_transformer)

epoxy_style_apply(.f = identity, ..., transformer = glue::identity_transformer)

epoxy_style_code(syntax = NULL, transformer = glue::identity_transformer)

epoxy_style_collapse(
  sep = ", ",
  last = sep,
  language = NULL,
  ...,
  transformer = glue::identity_transformer
)
}
\arguments{
\item{...}{A list of style functions, e.g. \code{epoxy_style_bold} or the name of
a style function, e.g. \code{"bold"}, or a call to a style function, e.g.
\code{epoxy_style_bold()}. \code{epoxy_style()} chains the style functions together,
applying the styles from left to right.

For example, \code{epoxy_style("bold", "collapse")} results in replaced strings
that are emboldened \emph{and then} collapsed, e.g. \verb{**a** and **b**}. On the
other hand, \code{epoxy_style("collapse", "bold")}  will collapse the vector
\emph{and then} embolden the entire string.

In \code{epoxy_style_apply()}, the \code{...} are passed to the underlying call the
underlying function call.

In \code{epoxy_style_collapse()}, the \code{...} are ignored.}

\item{syntax}{One of \code{"markdown"} (or \code{"md"}), \code{"html"}, or \code{"latex"}. The
default is chosen based on the engine of the chunk where the style function
is called, or according to the option \code{epoxy.engine}. Caution: invalid
options are silently ignored, falling back to markdown syntax.}

\item{before, after}{In \code{epoxy_style_wrap()}, the characters to be added
before and after variables in the template string.}

\item{transformer}{The transformer to apply to the replacement string. This
argument is used for chaining the transformer functions. By providing a
function to this argument you can apply an additional transformation after
the current transformation. In nearly all cases, you can let
\code{epoxy_style()} handle this for you. The chain ends when
\code{\link[glue:identity_transformer]{glue::identity_transformer()}} is used as the \code{transformer}.}

\item{.f}{A function, function name or \code{\link[purrr:map]{purrr::map()}}-style inline function.}

\item{sep, last}{The separator to use when joining the vector elements when
the expression ends with a \code{*}. Elements are separated by \code{sep}, except for
the last two elements, which use \code{last}.}

\item{language}{In \code{epoxy_style_collapse()}, \code{language} is passed to
\code{\link[and:and]{and::and()}} or \code{\link[and:and]{and::or()}} to choose the correct and/or phrase and spacing
for the \code{language}. By default, will follow the system language. See
\link[and:and_languages]{and::and_languages} for supported languages.}
}
\value{
A function of \code{text} and \code{envir} suitable for the \code{.transformer}
argument of \code{\link[glue:glue]{glue::glue()}}.
}
\description{
These transformers provide additional automatic formatting for the template
strings. They are designed to be used with the \code{.transformer} chunk option of
in \code{epoxy} chunks. You can use \code{epoxy_style()} to chain several transformers
together. \code{epoxy_style()} and individual \pkg{epoxy} style functions can be
used in \code{epoxy}, \code{epoxy_html} and \code{epoxy_latex} chunks and will choose the
correct syntax for each.
}
\section{Functions}{
\itemize{
\item \code{epoxy_style_wrap()}: Wrap variables

\item \code{epoxy_style_bold()}: Embolden variables using \verb{**} in markdown, \verb{<strong>}
in HTML, or \verb{\\textbf\{\}} in LaTeX

\item \code{epoxy_style_italic()}: Italicize variables using \verb{_} in markdown, \verb{<em>} in
HTML, or \verb{\\emph\{\}} in LaTeX

\item \code{epoxy_style_apply()}: Apply a function to all replacement expressions

\item \code{epoxy_style_code()}: Code format variables using \verb{``} in markdown,
\verb{<code>} in HTML, or \verb{\\texttt\{\}} in LaTeX

\item \code{epoxy_style_collapse()}: Collapse vector variables

}}
\section{Output-specific styling}{

The \code{epoxy_style_} functions will attempt to use the correct syntax for
styling the replacement text for markdown, HTML and LaTeX. This choice is
driven by the chunk engine where the styling function is used. The \code{epoxy}
engine corresponds to markdown, \code{epoxy_html} to HTML, and \code{epoxy_latex} to
LaTeX.

Automatic syntax selection only works when the epoxy style functions are used
with epoxy knitr engines and during the knitr rendering process. When
used outside of this context, you can choose the desired syntax by setting
the \code{syntax} to one of \code{"markdown"}, \code{"html"} or \code{"latex"}.
}

\seealso{
Other epoxy-style glue transformers: 
\code{\link{epoxy_style_format}()},
\code{\link{epoxy_style_html}()}
}
\concept{epoxy-style glue transformers}
