% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{epoxyHTML}
\alias{epoxyHTML}
\title{Epoxy HTML Output for Shiny}
\usage{
epoxyHTML(
  .id,
  ...,
  .class = NULL,
  .class_item = NULL,
  .container = "div",
  .container_item = "span",
  .placeholder = "",
  .sep = "",
  .open = "{{",
  .close = "}}",
  .na = "",
  .trim = FALSE,
  .watch = NULL
)
}
\arguments{
\item{.id}{The output id}

\item{...}{UI elements or text (that will be treated as HTML), containing
template variables. Use named values to provide initial placeholder values.}

\item{.class}{Classes added to the output div, in addition to \code{.epoxy-html}}

\item{.class_item}{Classes added to the \code{.container} wrapping each template
variable.}

\item{.container}{The name of the HTML element to be used for the output
element, by default \code{"div"}.}

\item{.container_item}{The name of the HTML element to be used for each template item,
by default \code{"span"}.}

\item{.placeholder}{Default placeholder if a template variable placeholder
isn't provided.}

\item{.sep}{[\code{character(1)}: \sQuote{""}]\cr Separator used to separate elements.}

\item{.open}{Opening template variable delimiter}

\item{.close}{Closing template variable delimiter}

\item{.na}{[\code{character(1)}: \sQuote{NA}]\cr Value to replace NA values
with. If \code{NULL} missing values are propagated, that is an \code{NA} result will
cause \code{NA} output. Otherwise the value is replaced by the value of \code{.na}.}

\item{.trim}{[\code{logical(1)}: \sQuote{TRUE}]\cr Whether to trim the input
template with \code{trim()} or not.}

\item{.watch}{The names of placeholders that should be watched as inputs.
Clicking on these elements triggers \verb{input$<.id>_<placeholder>_clicked}.}
}
\value{
An HTML object.
}
\description{
Expermimental. An glue-like output for Shiny. \code{epoxyHTML()} lets you use
placeholders in your HTML such as \code{"{{height}}"}, that are provided values
from the server by giving \code{renderEpoxyHTML()} a \code{height} value.
}
\section{HTML Markup}{
 By default, placeholders are inserted into a \verb{<span>}
element in your UI, with the classes specified in \code{.class_item}.

\code{epoxyHTML()} also supports an HTML markup syntax similar to
\href{https://pughtml.com/what-is-pug-html}{pug} (an HTML preprocessor). With the
markup syntax, \code{"{{h3.example.basic\%basic-three demo}}"} creates a \code{demo}
placeholder inside an \verb{<h3 id="basic-three" class="example basic"></h3>} tag.

The placeholder template string follows the pattern \verb{\{\{<markup> <name>\}\}}.
The markup syntax comes first, separated from the placeholder name by a space.
The HTML element is first, followed by classes prefixed with \code{.} or and ID
prefixed with \verb{\%}. Note that due to the way that \code{\link[glue:glue]{glue::glue()}} parses the
template string, an \verb{\%} is needed instead of \verb{#}. The template markup can
contain only one element and one ID, but many classes can be specified.
}

\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  h2("epoxyHTML demo"),
  epoxy:::epoxyHTML(
    'test',
    fluidRow(
      tags$div(
        class = "col-xs-4",
        selectInput(
          inputId = "thing",
          label = "What is this {{color}} thing?",
          choices = c("apple", "banana", "coconut", "dolphin")
        )
      ),
      tags$div(
        class = "col-xs-4",
        selectInput(
          inputId = "color",
          label = "What color is the {{thing}}?",
          c("red", "blue", "black", "green", "yellow")
        )
      ),
      tags$div(
        class = "col-xs-4",
        sliderInput(
          inputId = "height",
          label = "How tall is the {{color}} {{thing}}?",
          value = 5,
          min = 0,
          max = 10,
          step = 0.1,
          post = "ft"
        )
      )
    ),
    tags$p(class = "big", "The {{color}} {{thing}} is {{height}} feet tall."),
    thing = "THING",
    color = "COLOR",
    height = "HEIGHT",
    .class_item = "inner"
  ),
  tags$style(HTML(
    '.big { font-size: 1.5em; }
    .inner:not(.epoxy-item__placeholder) { background-color: rgba(254, 233, 105, 0.5)}
    .epoxy-item__placeholder { color: #999999; }'
  ))
)

server <- function(input, output, session) {
  output$test <- epoxy:::renderEpoxyHTML(
    thing = input$thing,
    color = input$color,
    height = input$height
  )
}

shinyApp(ui, server)
}

}
\seealso{
renderEpoxyHTML
}
